Config: add cache. CreateWallet: change code
