+++ src/front/shared/pages/CreateWallet/Steps/SecondStep.js
@@ -3,11 +3,14 @@ import React, { useState, useEffect, useRef } from 'react'
 import CSSModules from 'react-css-modules'
 import styles from '../CreateWallet.scss'
 
+import { connect } from 'redaction'
 import reducers from 'redux/core/reducers'
 
+import ReactTooltip from 'react-tooltip'
 import { FormattedMessage, injectIntl } from 'react-intl'
 import { isMobile } from 'react-device-detect'
 
+import config from 'app-config'
 import actions from 'redux/actions'
 import { firebase, constants, stats } from 'helpers'
 import firestore from 'helpers/firebase/firestore'
@@ -24,7 +27,17 @@ import Cupture, {
 
 
 const SecondStep = (props) => {
-  const { intl: { locale }, onClick, currencies, error, setError, forcedCurrencyData, btcData, etcClick } = props
+  const { 
+    intl: { locale }, 
+    onClick, 
+    currencies, 
+    error, 
+    setError, 
+    forcedCurrencyData, 
+    btcData,
+    etcClick,
+    ethData 
+  } = props
 
   const _protection = {
     nothing: {
@@ -360,7 +373,6 @@ const SecondStep = (props) => {
                     </div>
                   </div>
                 </div>
-
               )
             })}
           </div>
